@@ -14,7 +14,7 @@
     "@reduxjs/toolkit": "^1.6.2",
     "@tootallnate/once": "^3.0.0",
     "@types/concaveman": "^1.1.3",
-    "@types/d3": "^5.7.2",
+    "@types/d3v5": "npm:@types/d3@^5.7.2",
     "@types/jest": "^26.0.15",
     "@types/node": "^14.17.32",
     "@types/offscreencanvas": "^2019.7.0",
@@ -39,9 +39,8 @@
     "jstat": "^1.9.4",
     "libtess": "^1.2.2",
     "lodash": "^4.17.20",
-    "material-ui-search-bar": "^1.0.0",
     "papaparse": "^5.3.1",
-    "react-data-grid": "^7.0.0-beta.11",
+    "react-data-grid": "7.0.0-beta.11",
     "react-promise-tracker": "^2.1.0",
     "react-rnd": "^10.2.2",
     "react-split": "^2.0.9",
@@ -51,6 +50,7 @@
     "redux-thunk": "^2.4.0",
     "regenerator-runtime": "^0.13.3",
     "resize-observer": "^1.0.0",
+    "sass": "^1.55.0",
     "serve-dist": "^1.0.11",
     "sigma-webpack": "^1.2.0-1",
     "three": "^0.117.1",
@@ -60,12 +60,22 @@
     "url-loader": "^4.1.1",
     "vega": "5.17.0",
     "vega-lite": "4.16.8",
-    "vega-typings": "0.19.1",
-    "worker-loader": "^3.0.8"
+    "vega-typings": "0.19.1"
   },
   "devDependencies": {
+    "@babel/core": "^7.10.2",
+    "@babel/preset-env": "^7.10.2",
+    "@babel/preset-react": "^7.8.3",
+    "@babel/preset-typescript": "^7.8.3",
+    "@emotion/react": "^11.5.0",
+    "@emotion/styled": "^11.3.0",
+    "@mui/material": "^5.0.6",
+    "@mui/styles": "^5.0.2",
+    "@svgr/webpack": "^5.5.0",
     "@typescript-eslint/eslint-plugin": "^5.9.0",
     "@typescript-eslint/parser": "^5.9.0",
+    "babel-loader": "^8.2.3",
+    "css-loader": "^3.4.2",
     "eslint": "^8.7.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^16.1.0",
@@ -76,34 +86,20 @@
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.28.0",
     "eslint-plugin-react-hooks": "^4.3.0",
-    "prettier": "^2.5.1",
-    "@babel/core": "^7.10.2",
-    "@babel/preset-env": "^7.10.2",
-    "@babel/preset-react": "^7.8.3",
-    "@babel/preset-typescript": "^7.8.3",
-    "@emotion/react": "^11.5.0",
-    "@emotion/styled": "^11.3.0",
-    "@mui/material": "^5.0.6",
-    "@mui/styles": "^5.0.2",
-    "@svgr/webpack": "^5.5.0",
-    "babel-loader": "^8.2.3",
-    "css-loader": "^3.4.2",
     "jest": "^26.6.3",
-    "node-sass": "^4.13.1",
+    "prettier": "^2.5.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.6",
     "redux": "^4.1.2",
     "sass-loader": "^8.0.2",
-    "shader-loader": "^1.3.1",
     "source-map-loader": "^0.2.4",
     "style-loader": "^1.1.3",
     "svg-url-loader": "^7.1.1",
     "ts-loader": "^8.2.0",
     "typescript": "^4.7.4",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.10",
-    "webpack-glsl-loader": "^1.0.1"
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0"
   },
   "peerDependencies": {
     "@emotion/react": "^11.5.0",
@@ -116,7 +112,7 @@
     "redux": "^4.0.5"
   },
   "scripts": {
-    "webpack:dev": "npx webpack --config webpack.config.js",
+    "webpack:dev": "npx webpack --config webpack.config.js --stats-children",
     "test": "jest",
     "lint": "eslint src/**/*.ts{,x}",
     "lint:fix": "eslint --fix src/**/*.ts{,x}"
