@@ -2,10 +2,10 @@
 /* eslint-disable react/no-unstable-nested-components */
 import { Checkbox, TableCell } from '@mui/material';
 import { connect, ConnectedProps } from 'react-redux';
-import React = require('react');
+import * as React from 'react';
 import { AutoSizer, Column, Table, defaultTableRowRenderer } from 'react-virtualized';
 import clsx from 'clsx';
-import clone = require('fast-clone');
+import clone from 'fast-clone';
 import { withStyles } from '@mui/styles';
 import { setProjectionColumns } from '../../Ducks/ProjectionColumnsDuck';
 import { setProjectionOpenAction } from '../../Ducks/ProjectionOpenDuck';
