@@ -1,14 +1,12 @@
 /* eslint-disable spaced-comment */
 /* eslint-disable @typescript-eslint/no-this-alias */
 /* eslint-disable new-cap */
-import Graph = require('graphology');
+import * as Graph from 'graphology';
 import { DataLine } from '../../../model/DataLine';
 import { Dataset } from '../../../model/Dataset';
 import { EmbeddingController } from './EmbeddingController';
 
 import { IVector } from '../../../model/Vector';
-
-import forceatlas2Worker from '../../workers/embeddings/forceatlas2.worker';
 
 export class ForceAtlas2EmbeddingController extends EmbeddingController {
   nodes: any;
@@ -69,7 +67,7 @@ export class ForceAtlas2EmbeddingController extends EmbeddingController {
       graph.addEdge(edge.source, edge.destination);
     });
 
-    this.worker = new forceatlas2Worker();
+    this.worker = new Worker(new URL('../../workers/embeddings/forceatlas2.worker', import.meta.url));
 
     const self = this;
     this.worker.onmessage = function (e) {
