@@ -1,4 +1,4 @@
-import React = require('react');
+import * as React from 'react';
 import { Autocomplete, createFilterOptions, TextField } from '@mui/material';
 
 interface SelectFeatureComponentProps {
