@@ -2,8 +2,8 @@
 import * as d3v5 from 'd3v5';
 import { Scene, Vector2, Vector3 } from 'three';
 import { EntityId } from '@reduxjs/toolkit';
-import THREE = require('three');
-import React = require('react');
+import * as THREE from 'three';
+import * as React from 'react';
 import { connect, ConnectedProps } from 'react-redux';
 import { Typography } from '@mui/material';
 import { IVector } from '../../model/Vector';
@@ -17,7 +17,6 @@ import { IBook } from '../../model/Book';
 import * as nt from '../NumTs/NumTs';
 
 import { GroupVisualizationMode } from '../Ducks/GroupVisualizationMode';
-import TessyWorker from '../workers/tessy.worker';
 import { ViewTransformType } from '../Ducks';
 import { SchemeColor } from '../Utility/Colors/SchemeColor';
 import { AStorytelling } from '../Ducks/StoriesDuck copy';
@@ -141,7 +140,7 @@ export const MultivariateClustering = connector(
       this.trail.create();
       this.scene.add(this.trail.mesh);
 
-      this.triangulationWorker = new TessyWorker();
+      this.triangulationWorker = new Worker(new URL('../workers/tessy.worker', import.meta.url));
     }
 
     componentDidMount() {
@@ -219,8 +218,14 @@ export const MultivariateClustering = connector(
         }
       }
 
-      if (prevProps.workspace !== this.props.workspace) {
-        this.updatePositions(this.props.viewTransform.zoom);
+      if (prevProps.workspace !== this.props.workspace && this.props.workspace) {
+        // TODO: check performance implications
+        // this.updatePositions(this.props.viewTransform.zoom);
+        //this.destroy();
+        //this.disposeTriangulatedMesh();
+
+        //this.create();
+        //this.createTriangulatedMesh();
       }
 
       if (this.props.onInvalidate) {
@@ -478,10 +483,10 @@ export const MultivariateClustering = connector(
       });
       if (this.clusterVis) {
         this.clusterVis.lineMeshes.forEach((mesh) => {
-          mesh.visible = false;
+          //mesh.visible = false;
         });
         this.clusterVis.clusterMeshes?.forEach((mesh) => {
-          mesh.visible = false;
+          //mesh.visible = false;
         });
       }
     }
@@ -556,11 +561,13 @@ export const MultivariateClustering = connector(
 
           const contours = d3v5
             .contourDensity()
-            .x((d) => xAxis(d.x))
-            .y((d) => yAxis(d.y))
+            .x((d) => xAxis(d[0]))
+            .y((d) => yAxis(d[1]))
             .bandwidth(10)
             .thresholds(10)
-            .size([100, bounds.width === 0 ? 1 : Math.floor(100 * (bounds.height / bounds.width))])(cluster.indices.map((i) => this.props.workspace[i]));
+            .size([100, bounds.width === 0 ? 1 : Math.floor(100 * (bounds.height / bounds.width))])(
+            cluster.indices.map((i) => [this.props.workspace[i].x, this.props.workspace[i].y]),
+          );
 
           const clusterObject = this.clusterObjects.find((e) => activeStory.clusters.entities[e.cluster].label === cluster.label);
 
