@@ -2,7 +2,7 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 import * as THREE from 'three';
 import { BufferAttribute } from 'three';
-import { valueInRange } from './UtilityFunctions';
+import { getMinMaxOfChannel, valueInRange } from './UtilityFunctions';
 import { ContinuousMapping, DiscreteMapping, Mapping } from '../Utility/Colors/Mapping';
 import { DataLine } from '../../model/DataLine';
 import { IVector } from '../../model/Vector';
@@ -16,10 +16,12 @@ import { createLinearRangeScaler } from '../Utility/ScalingAndAxes';
 import { Shapes } from './Shapes';
 import { IStorytelling, AStorytelling } from '../Ducks/StoriesDuck copy';
 
-const fragmentShader = require('../../shaders/fragment.glsl');
-const vertexShader = require('../../shaders/vertex.glsl');
+// @ts-ignore
+import fragmentShader from '../../shaders/fragment.glsl?raw';
+// @ts-ignore
+import vertexShader from '../../shaders/vertex.glsl?raw';
 
-const override = require('./meshline');
+import override from './meshline';
 
 export function imageFromShape(value) {
   switch (value) {
@@ -576,7 +578,7 @@ export class PointVisualization {
   /**
    * @param {*} category a feature to select the shape for
    */
-  shapeCat(category) {
+  setShapeByChannel(category) {
     const type = this.mesh.geometry.attributes.type as BufferAttribute;
 
     // default shapes used
@@ -619,7 +621,7 @@ export class PointVisualization {
     type.needsUpdate = true;
   }
 
-  colorCat(category, scale) {
+  setColorByChannel(category, scale) {
     this.colorAttribute = category;
 
     if (category == null) {
@@ -647,50 +649,32 @@ export class PointVisualization {
     }
   }
 
-  transparencyCat(category, range) {
+  setBrightnessByChannel(channel, range) {
     // var color = this.mesh.geometry.attributes.customColor.array
 
-    if (category == null) {
+    if (channel == null) {
       // default transparency
       this.vectors.forEach((sample) => {
         sample.__meta__.brightness = range[0];
       });
-    } else if (category.type === 'sequential') {
+    } else if (channel.type === 'sequential') {
       if (this.dataset.isSequential) {
         this.segments.forEach((segment) => {
-          let min = null;
-          let max = null;
-          if (this.dataset.columns[category.key].range) {
-            min = this.dataset.columns[category.key].range.min;
-            max = this.dataset.columns[category.key].range.max;
-          } else {
-            const filtered = segment.vectors.map((vector) => vector[category.key]);
-            max = Math.max(...filtered);
-            min = Math.min(...filtered);
-          }
+          const { min, max } = getMinMaxOfChannel(this.dataset, channel.key, segment);
 
           const scaler = createLinearRangeScaler(range, min, max);
 
           segment.vectors.forEach((vector) => {
-            vector.__meta__.brightness = scaler(vector[category.key]);
+            vector.__meta__.brightness = scaler(vector[channel.key]);
           });
         });
       } else {
-        let min = null;
-        let max = null;
-        if (this.dataset.columns[category.key].range) {
-          min = this.dataset.columns[category.key].range.min;
-          max = this.dataset.columns[category.key].range.max;
-        } else {
-          const filtered = this.vectors.map((vector) => vector[category.key]);
-          max = Math.max(...filtered);
-          min = Math.min(...filtered);
-        }
+        const { min, max } = getMinMaxOfChannel(this.dataset, channel.key);
 
         const scaler = createLinearRangeScaler(range, min, max);
 
         this.vectors.forEach((vector) => {
-          vector.__meta__.brightness = scaler(vector[category.key]);
+          vector.__meta__.brightness = scaler(vector[channel.key]);
         });
       }
     }
@@ -708,16 +692,7 @@ export class PointVisualization {
         if (this.dataset.isSequential) {
           // dataset with lines, we have segments
           this.segments.forEach((segment) => {
-            let min = null;
-            let max = null;
-            if (this.dataset.columns[category.key].range) {
-              min = this.dataset.columns[category.key].range.min;
-              max = this.dataset.columns[category.key].range.max;
-            } else {
-              const filtered = segment.vectors.map((vector) => vector[category.key]);
-              max = Math.max(...filtered);
-              min = Math.min(...filtered);
-            }
+            const { min, max } = getMinMaxOfChannel(this.dataset, category.key, segment);
 
             const sizeScaler = createLinearRangeScaler(range, min, max);
 
@@ -727,16 +702,7 @@ export class PointVisualization {
           });
         } else {
           // for state based data, min and max is based on whole dataset
-          let min = null;
-          let max = null;
-          if (this.dataset.columns[category.key].range) {
-            min = this.dataset.columns[category.key].range.min;
-            max = this.dataset.columns[category.key].range.max;
-          } else {
-            const filtered = this.vectors.map((vector) => vector[category.key]);
-            max = Math.max(...filtered);
-            min = Math.min(...filtered);
-          }
+          const { min, max } = getMinMaxOfChannel(this.dataset, category.key);
 
           const sizeScaler = createLinearRangeScaler(range, min, max);
 
@@ -911,8 +877,8 @@ export class PointVisualization {
   /**
    * Updates the zoom level.
    */
-  zoom(zoom) {
-    this.mesh.material.uniforms.zoom.value = zoom * this.dataset.bounds.scaleFactor;
+  zoom(zoom, projection) {
+    this.mesh.material.uniforms.zoom.value = zoom * projection.bounds.scaleFactor;
   }
 
   /**
