@@ -1,4 +1,4 @@
-import React = require('react');
+import * as React from 'react';
 import { Typography, Slider } from '@mui/material';
 import { connect, ConnectedProps } from 'react-redux';
 import { setDifferenceThreshold } from '../../../components/Ducks/DifferenceThresholdDuck';
