@@ -1,4 +1,4 @@
-import React = require('react');
+import * as React from 'react';
 import { createClassFromSpec, VisualizationSpec, VegaLite } from 'react-vega';
 import { VegaLiteProps } from 'react-vega/lib/VegaLite';
 
