@@ -68,7 +68,7 @@ export type DetailViewSpec = {
 export type TabSpec = {
   name: string;
   tab: JSX.Element | ((props: any) => JSX.Element) | ConnectedComponent<any, any>;
-  icon: () => JSX.Element;
+  icon: JSX.Element | (() => JSX.Element);
   title: string;
   description: string;
 };
