@@ -46,10 +46,8 @@ function WebView({
       ? multiples.projections.entities[value.attributes.workspace]
       : value.attributes.workspace;
 
-  const [positions, setPositions] = React.useState<IPosition[]>();
-
-  React.useEffect(() => {
-    setPositions(selectPositions(dataset, projection));
+  const positions = React.useMemo(() => {
+    return selectPositions(dataset, projection);
   }, [projection.xChannel, projection.yChannel, dataset, projection]);
 
   const active = value.id === multiples.active;
@@ -88,12 +86,10 @@ function WebView({
             {projection.metadata?.method}
           </Typography>
 
-          {projection.name ? (
-            <Typography variant="body1" component="span">
-              {` - "${projection.name}"`}
-              {projection.xChannel || projection.yChannel ? ` [x: ${projection.xChannel}, y: ${projection.yChannel}]` : null}
-            </Typography>
-          ) : null}
+          <Typography variant="body1" component="span">
+            {projection.name ? ` - "${projection.name}"` : ''}
+            {projection.xChannel || projection.yChannel ? ` [x: ${projection.xChannel}, y: ${projection.yChannel}]` : null}
+          </Typography>
         </div>
 
         <div style={{ display: 'flex' }}>
@@ -113,7 +109,17 @@ function WebView({
         </div>
       </div>
       <div style={{ flexGrow: 1 }}>
-        {positions ? <WebGLView overrideComponents={overrideComponents} multipleId={id} {...value.attributes} workspace={positions} /> : null}
+        {positions && projection ? (
+          <WebGLView
+            overrideComponents={overrideComponents}
+            multipleId={id}
+            {...value.attributes}
+            projection={projection}
+            workspace={positions}
+            xChannel={projection.xChannel}
+            yChannel={projection.yChannel}
+          />
+        ) : null}
       </div>
     </div>
   );
