@@ -43,10 +43,10 @@ import { DataLine } from '../../model/DataLine';
 import { ObjectTypes } from '../../model/ObjectType';
 import { ComponentConfig } from '../../BaseConfig';
 import { ANormalized } from '../Utility/NormalizedState';
-import { StoriesActions, AStorytelling } from '../Ducks/StoriesDuck copy';
+import { StoriesActions, AStorytelling, bookAdapter } from '../Ducks/StoriesDuck copy';
 import { Mapping } from '../Utility';
 import { ViewActions, SingleMultipleAttributes, ViewSelector } from '../Ducks/ViewDuck';
-import { IPosition } from '../../model';
+import { AProjection, IPosition, IProjection } from '../../model';
 
 type ViewState = {
   camera: Camera;
@@ -98,6 +98,7 @@ type Props = PropsFromRedux & {
   // The id for this visualization, important for selecting the correct slice
   multipleId: EntityId;
   workspace: IPosition[];
+  projection: IProjection;
 } & Omit<SingleMultipleAttributes, 'workspace'>;
 
 const UPDATER = 'scatter';
@@ -128,11 +129,11 @@ export const WebGLView = connector(
 
     currentHover: TypedObject;
 
-    camera: any;
+    camera: THREE.OrthographicCamera;
 
     vectors: IVector[];
 
-    renderer: any;
+    renderer: THREE.WebGLRenderer;
 
     lines: LineVisualization;
 
@@ -377,14 +378,16 @@ export const WebGLView = connector(
 
       const normalized = normalizeWheel(event);
 
+      // AProjection.calculateBounds(this.props.dataset, this.props.projection.xChannel, this.props.projection.yChannel, this.props.workspace);
+
       // Store world position under mouse
       const bounds = this.containerRef.current.getBoundingClientRect();
       const worldBefore = CameraTransformations.screenToWorld({ x: event.clientX - bounds.left, y: event.clientY - bounds.top }, this.createTransform());
       const screenBefore = this.relativeMousePosition(event);
 
-      const newZoom = this.camera.zoom - (normalized.pixelY * 0.013) / this.props.dataset.bounds.scaleFactor;
-      if (newZoom < 1.0 / this.props.dataset.bounds.scaleFactor) {
-        this.camera.zoom = 1.0 / this.props.dataset.bounds.scaleFactor;
+      const newZoom = this.camera.zoom - (normalized.pixelY * 0.013) / this.props.projection.bounds.scaleFactor;
+      if (newZoom < 1.0 / this.props.projection.bounds.scaleFactor) {
+        this.camera.zoom = 1.0 / this.props.projection.bounds.scaleFactor;
       } else {
         this.camera.zoom = newZoom;
       }
@@ -393,7 +396,7 @@ export const WebGLView = connector(
       this.restoreCamera(worldBefore, screenBefore);
 
       // Adjust mesh zoom levels
-      this.particles.zoom(this.camera.zoom);
+      this.particles.zoom(this.camera.zoom, this.props.projection);
 
       if (this.props.dataset.isSequential) {
         this.lines.setZoom(this.camera.zoom);
@@ -721,10 +724,13 @@ export const WebGLView = connector(
       this.renderer.setSize(this.getWidth(), this.getHeight());
       this.renderer.setClearColor(0x000000, 0);
       this.renderer.sortObjects = false;
+      this.renderer.localClippingEnabled = false;
 
-      this.camera = new THREE.OrthographicCamera(this.getWidth() / -2, this.getWidth() / 2, this.getHeight() / 2, this.getHeight() / -2, 1, 1000);
+      this.camera = new THREE.OrthographicCamera(this.getWidth() / -2, this.getWidth() / 2, this.getHeight() / 2, this.getHeight() / -2, 0, 1000);
       this.camera.position.z = 1;
       this.camera.lookAt(new THREE.Vector3(0, 0, 0));
+      this.camera.near = 0;
+      this.camera.far = 100000;
 
       // this.camera.position.x = this.props.viewTransform.centerX
       // this.camera.position.y = this.props.viewTransform.centerY
@@ -746,10 +752,20 @@ export const WebGLView = connector(
       this.lineColorScheme = lineColorScheme;
       this.vectorMapping = vectorMapping;
 
+      const { zoom, x, y } = getDefaultZoom(
+        dataset,
+        this.getWidth(),
+        this.getHeight(),
+        this.props.projection.xChannel,
+        this.props.projection.yChannel,
+        this.props.workspace,
+      );
+
       // Update camera zoom to fit the problem
-      this.camera.zoom = getDefaultZoom(this.props.dataset.vectors, this.getWidth(), this.getHeight());
-      this.camera.position.x = 0.0;
-      this.camera.position.y = 0.0;
+      this.camera.zoom = zoom;
+      this.camera.position.x = x;
+      this.camera.position.y = y;
+      this.camera.position.z = 10;
       this.camera.updateProjectionMatrix();
 
       this.props.setViewTransform(this.camera, this.getWidth(), this.getHeight());
@@ -775,9 +791,10 @@ export const WebGLView = connector(
         this.segments,
       );
       this.particles.createMesh(this.props.dataset.vectors, this.segments, () => {
+        this.updateXY();
         this.requestRender();
       });
-      this.particles.zoom(this.camera.zoom);
+      this.particles.zoom(this.camera.zoom, this.props.projection);
       this.particles.update();
 
       // this.scene.add(this.particles.mesh);
@@ -853,12 +870,22 @@ export const WebGLView = connector(
           this.lines.updatePosition(this.props.workspace);
         }
 
-        ADataset.calculateBounds(this.props.dataset);
+        // AProjection.calculateBounds(this.props.dataset, this.props.projection.xChannel, this.props.projection.yChannel, this.props.workspace);
+        // ADataset.calculateBounds(this.props.dataset, this.props.projection.xChannel, this.props.projection.yChannel, this.props.workspace);
 
-        this.camera.zoom = getDefaultZoom(this.props.dataset.vectors, this.getWidth(), this.getHeight());
+        const { zoom, x, y } = getDefaultZoom(
+          this.props.dataset,
+          this.getWidth(),
+          this.getHeight(),
+          this.props.projection.xChannel,
+          this.props.projection.yChannel,
+          this.props.workspace,
+        );
 
-        this.camera.position.x = 0.0;
-        this.camera.position.y = 0.0;
+        this.camera.zoom = zoom;
+
+        this.camera.position.x = x;
+        this.camera.position.y = y;
 
         this.camera.updateProjectionMatrix();
 
@@ -1062,13 +1089,15 @@ export const WebGLView = connector(
           );
           this.props.setPointColorMapping(this.props.multipleId, mapping);
 
-          this.particles.colorCat(this.props.channelColor, mapping);
+          this.particles.setColorByChannel(this.props.channelColor, mapping);
         }
       }
       this.startRendering();
     }
 
     componentDidUpdate(prevProps: Props, prevState) {
+      // console.log(this.props.workspace);
+      // console.log(this.props.xChannel);
       if (prevProps.dataset !== this.props.dataset) {
         this.createVisualization(this.props.dataset, mappingFromScale({ type: 'categorical', palette: 'dark2' }, { key: 'algo' }, this.props.dataset), null);
       }
@@ -1085,10 +1114,10 @@ export const WebGLView = connector(
             this.props.dataset,
           );
           this.props.setPointColorMapping(this.props.multipleId, mapping);
-          this.particles.colorCat(this.props.channelColor, mapping);
+          this.particles.setColorByChannel(this.props.channelColor, mapping);
         } else {
           this.props.setPointColorMapping(this.props.multipleId, null);
-          this.particles?.colorCat(null, null);
+          this.particles?.setColorByChannel(null, null);
         }
       }
 
@@ -1121,7 +1150,7 @@ export const WebGLView = connector(
         prevProps.globalPointBrightness !== this.props.globalPointBrightness ||
         (prevProps.channelBrightness !== this.props.channelBrightness && this.particles)
       ) {
-        this.particles.transparencyCat(this.props.channelBrightness, this.props.globalPointBrightness);
+        this.particles.setBrightnessByChannel(this.props.channelBrightness, this.props.globalPointBrightness);
         this.particles.updateColor();
         this.particles.update();
       }
@@ -1195,7 +1224,7 @@ export const WebGLView = connector(
 
       if (prevProps.vectorByShape !== this.props.vectorByShape && this.particles) {
         this.filterPoints({ star: true, cross: true, circle: true, square: true });
-        this.particles.shapeCat(this.props.vectorByShape);
+        this.particles.setShapeByChannel(this.props.vectorByShape);
       }
 
       if (prevProps.pointDisplay.checkedShapes !== this.props.pointDisplay.checkedShapes && this.particles) {
@@ -1384,6 +1413,7 @@ export const WebGLView = connector(
                         entities: {},
                         ids: [],
                       },
+                      metadata: { method: 'custom' },
                     };
 
                     this.props.addStory(story, true);
